# product-sale-prediction-AI/resources/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Product Prediction API
  description: |
    Intelligent E-commerce API with vector search, sales analytics, and prediction capabilities.
    This API provides advanced features for product management, search, analytics, and sales predictions.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.example.com
    description: Production Server
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: api-key
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: "P123456"
          description: Unique product identifier
        name:
          type: string
          example: "Wireless Bluetooth Headphones"
          description: Product name
        category:
          type: string
          example: "Electronics"
          description: Product category
        brand:
          type: string
          example: "AudioTech"
          description: Product brand
        price:
          type: number
          format: double
          example: 99.99
          description: Product price
        createTimestamp:
          type: string
          format: date-time
          example: "2023-05-15T10:30:00"
          description: Product creation timestamp
        description:
          type: string
          example: "High-quality wireless headphones with noise cancellation"
          description: Detailed product description
    Order:
      type: object
      properties:
        orderId:
          type: string
          example: "ORD98765"
          description: Unique order identifier
        productId:
          type: string
          example: "P123456"
          description: Product identifier
        buyerId:
          type: string
          example: "USER4321"
          description: Customer who made the purchase
        sellerId:
          type: string
          example: "SELLER789"
          description: Seller who sold the product
        unitPrice:
          type: number
          format: double
          example: 99.99
          description: Price per unit
        quantity:
          type: integer
          example: 2
          description: Number of units ordered
        totalPrice:
          type: number
          format: double
          example: 199.98
          description: Total order price
        timestamp:
          type: string
          format: date-time
          example: "2023-07-20T14:25:00"
          description: Order timestamp
...
paths:
  /v1/products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve all products, optionally filtered by category and/or seller ID
      operationId: getAllProducts
      parameters:
        - name: category
          in: query
          description: Filter by product category
          required: false
          schema:
            type: string
          example: "Electronics"
        - name: sellerId
          in: query
          description: Filter by seller ID
          required: false
          schema:
            type: string
          example: "SELLER789"
      responses:
        '200':
          description: Successfully retrieved products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  count:
                    type: integer
                    example: 25
                  category:
                    type: string
                    example: "Electronics"
                  sellerId:
                    type: string
                    example: "SELLER789"
  /v1/product/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a product by its unique identifier
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: string
          example: "P123456"
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
...
  /v1/orders:
    get:
      tags:
        - Orders
      summary: Get recent orders
      description: Retrieve recent orders with pagination support (max 100 items per page)
      operationId: getRecentOrders
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            default: 0
          example: 0
        - name: size
          in: query
          description: Page size (max 100)
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          example: 20
      responses:
        '200':
          description: Successfully retrieved orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  currentPage:
                    type: integer
                    example: 0
                  totalItems:
                    type: integer
                    example: 250
                  totalPages:
                    type: integer
                    example: 13
...
  /api/mcp/products:
    get:
      tags:
        - MCP
      summary: Get all products (MCP)
      description: Retrieve all products with optional filtering, formatted for LLM consumption
      operationId: mcpGetAllProducts
      parameters:
        - name: category
          in: query
          description: Filter by product category
          required: false
          schema:
            type: string
          example: "Electronics"
        - name: sellerId
          in: query
          description: Filter by seller ID
          required: false
          schema:
            type: string
          example: "SELLER789"
      responses:
        '200':
          description: Successfully retrieved products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
...